Imported upstream version '3.2.0' of 'opencv'

+++ opencv_contrib/xfeatures2d/src/pct_signatures.cpp
@@ -397,7 +397,7 @@ namespace cv
                 break;
             case REGULAR:
             {
-                int gridSize = (int)ceil(sqrt((float)count));
+                int gridSize = (int)ceil(sqrt(count));
                 const float step = 1.0f / gridSize;
                 const float halfStep = step / 2;
                 float x = halfStep;
